requirements.txt
africastalking
intasend-python
Flask
psycopg2